pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "awe-agentcheck"
version = "0.1.0"
description = "Multi-CLI agent orchestration and cross-review platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "langgraph>=0.2.48,<0.4.0",
  "opentelemetry-sdk>=1.26.0,<2.0.0",
  "opentelemetry-exporter-otlp>=1.26.0,<2.0.0",
  "httpx>=0.27.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "pytest-cov>=5.0.0,<7.0.0",
  "ruff>=0.6.0,<1.0.0",
  "mypy>=1.11.0,<2.0.0",
  "bandit>=1.7.0,<2.0.0"
]

[project.scripts]
awe-agentcheck = "awe_agentcheck.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
pretty = true
warn_unused_ignores = true
ignore_missing_imports = true
follow_imports = "silent"

[[tool.mypy.overrides]]
module = [
  "tests.*",
  "awe_agentcheck.adapters.*",
  "awe_agentcheck.api",
  "awe_agentcheck.automation",
  "awe_agentcheck.benchmark",
  "awe_agentcheck.cli",
  "awe_agentcheck.db",
  "awe_agentcheck.fusion",
  "awe_agentcheck.main",
  "awe_agentcheck.repository",
  "awe_agentcheck.service",
  "awe_agentcheck.service_layers.*",
  "awe_agentcheck.workflow",
]
ignore_errors = true
